+++ com.unity.formats.fbx/Editor/ConvertToNestedPrefab.cs
@@ -271,8 +271,13 @@ namespace UnityEditor.Formats.Fbx.Exporter
         }
 
 #if UNITY_2021_2_OR_NEWER
-        internal static List<Object> GetSceneReferences(int instanceId)
+        internal static List<Object> GetSceneReferences(Object obj)
         {
+#if UNITY_6000_2_OR_NEWER
+            var instanceId = obj.GetEntityId();
+#else
+            var instanceId = obj.GetInstanceID();
+#endif
             var query = $"h: ref={instanceId}";
 
             using (var searchContext = UnityEditor.Search.SearchService.CreateContext(query))
@@ -341,7 +346,7 @@ namespace UnityEditor.Formats.Fbx.Exporter
         /// <returns></returns>
         internal static List<GameObject> GetSceneReferencesToObject(Object obj)
         {
-            var instanceID = obj.GetInstanceID();
+            var instanceId = obj.GetInstanceID();
             var idFormat = "ref:{0}:";
 
             var sceneHierarchyWindow = SceneHierarchyWindow;
@@ -351,7 +356,7 @@ namespace UnityEditor.Formats.Fbx.Exporter
             var previousSearchFilter = sceneHierarchy.GetType().GetField("m_SearchFilter", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance).GetValue(sceneHierarchy);
 
             // Set the search filter to find all references in the scene to the given object
-            setSearchFilterMethod.Invoke(sceneHierarchyWindow, new object[] { string.Format(idFormat, instanceID), SearchableEditorWindow.SearchMode.All, true, false });
+            setSearchFilterMethod.Invoke(sceneHierarchyWindow, new object[] { string.Format(idFormat, instanceId), SearchableEditorWindow.SearchMode.All, true, false });
 
             // Get objects from list of instance IDs of currently visible objects
             var treeView = GetPropertyReflection(sceneHierarchy, "treeView", isPublic: false);
@@ -363,7 +368,11 @@ namespace UnityEditor.Formats.Fbx.Exporter
             foreach (var row in rows)
             {
                 var id = (int)GetPropertyReflection(row, "id", isPublic: true);
+#if UNITY_6000_2_OR_NEWER
+                var gameObject = EditorUtility.EntityIdToObject(id) as GameObject;
+#else
                 var gameObject = EditorUtility.InstanceIDToObject(id) as GameObject;
+#endif
                 if (gameObject)
                 {
                     sceneObjects.Add(gameObject);
@@ -619,8 +628,7 @@ namespace UnityEditor.Formats.Fbx.Exporter
 #if UNITY_2021_2_OR_NEWER
                 if (!isPreviewScene)
                 {
-                    var instanceID = t.gameObject.GetInstanceID();
-                    info.sceneObjectsWithReference = GetSceneReferences(instanceID);
+                    info.sceneObjectsWithReference = GetSceneReferences(t.gameObject);
                 }
 #endif // UNITY_2021_2_OR_NEWER